Restructured directories for source
