+++ include/internal/csv_utility.h
@@ -1,0 +1,35 @@
+#pragma once
+#include "constants.hpp"
+
+#include <string>
+
+namespace csv {
+    /** Returned by get_file_info() */
+    struct CSVFileInfo {
+        std::string filename;               /**< Filename */
+        std::vector<std::string> col_names; /**< CSV column names */
+        char delim;                         /**< Delimiting character */
+        RowCount n_rows;                    /**< Number of rows in a file */
+        int n_cols;                         /**< Number of columns in a CSV */
+    };
+
+    /** @name Shorthand Parsing Functions
+     *  @brief Convienience functions for parsing small strings
+     */
+     ///@{
+    CSVCollection operator ""_csv(const char*, size_t);
+    CSVCollection parse(const std::string& in, CSVFormat format = DEFAULT_CSV);
+    ///@}
+
+    /** @name Utility Functions */
+    ///@{
+    std::unordered_map<std::string, DataType> csv_data_types(const std::string&);
+    CSVFileInfo get_file_info(const std::string& filename);
+    CSVFormat guess_format(const std::string& filename);
+    std::vector<std::string> get_col_names(
+        const std::string& filename,
+        const CSVFormat format = GUESS_CSV);
+    int get_col_pos(const std::string filename, const std::string col_name,
+        const CSVFormat format = GUESS_CSV);
+    ///@}
+}
