+++ programs/csv_stats.cpp
@@ -1,4 +1,4 @@
-#include "csv_parser.hpp"
+#include "csv.hpp"
 
 int main(int argc, char** argv) {
     using namespace csv;
