+++ tests/test_csv_iterator.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "catch.hpp"
-#include "csv_parser.hpp"
+#include "csv.hpp"
 using namespace csv;
 
 //////////////////////
